requirements.txt
asgiref==3.8.1
dj-static==0.0.6
Django==5.0.6
django-admin-interface==0.28.6
django-ckeditor==6.7.1
django-colorfield==0.11.0
django-js-asset==2.2.0
django-richtextfield==1.6.2
pillow==10.3.0
python-slugify==8.0.4
sqlparse==0.5.0
static3==0.7.0
text-unidecode==1.3
typing_extensions==4.11.0
tzdata==2024.1
whitenoise==6.6.0
asgiref==3.8.1
diff-match-patch==20230430
dj-static==0.0.6
Django==5.0.6
django-admin-interface==0.28.6
django-ckeditor==6.7.1
django-colorfield==0.11.0
django-import-export==4.0.8
django-js-asset==2.2.0
django-richtextfield==1.6.2
et-xmlfile==1.1.0
Faker==25.8.0
numpy==2.0.0
openpyxl==3.1.4
pandas==2.2.2
pillow==10.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
static3==0.7.0
tablib==3.5.0
text-unidecode==1.3
typing_extensions==4.11.0
tzdata==2024.1
whitenoise==6.6.0
